refactor: confirm gates on every db: command; drop db:reset
- Every db:* command now prints DATABASE_URL and asks y/N before doing
  anything (skip with -- -y). Reduces the risk that an inline
  DATABASE_URL override slips into a prod operation.
- Drop db:reset entirely. Dev mirrors prod operationally: both are
  persistent, mutated only via db:push. Wiping a database is a one-off
  SQL operation, not a permanent script.
- Rename db:seed → db:seed-dev so the dev-only data scope is visible at
  the command line.
