+++ package.json
@@ -16,8 +16,7 @@
 		"test:e2e": "playwright test",
 		"test": "vitest run && playwright test",
 		"db:push": "pnpx tsx scripts/db/push.ts",
-		"db:reset": "pnpx tsx scripts/db/reset.ts",
-		"db:seed": "pnpx tsx scripts/db/seed.ts",
+		"db:seed-dev": "pnpx tsx scripts/db/seed-dev.ts",
 		"db:studio": "pnpx tsx scripts/db/studio.ts",
 		"db:tunnel": "pnpx tsx scripts/db/tunnel.ts",
 		"auth:schema": "pnpm dlx auth@latest generate --config src/lib/server/auth.ts --output src/lib/server/db/auth.schema.ts -y && pnpm format",
